Extract configuration to constant
Before, the implementation has configuration within it. This means that
if cheese prices change, we have to change the code. If prices never or
rarely change, it might just be practical to leave this code as is.

Extracting it makes it easier to adjust cheese prices in our pricer as
the seasonal prices change. It also makes it easier to see what the
prices are since they are not hidden in the implementation.