Decouple pricer tests from price config
The tests for CheesePricer#pricer is no longer isolated to test only the
behavior implemented within it. It still has knowledge of the constants
that we've extracted and the test results are dependent on a specific
configuration.

To adapt more nimbly to changes in cheese prices, we can make the tests
agnostic to cheese prices. We do this here by stubbing PRICES so we can
test against a known, static configuration.